code/src/api/requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dateparser==1.2.1
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
GoogleNews==1.6.15
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==4.9.4
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.2
safetensors==0.5.3
sec-api==1.0.31
six==1.17.0
soupsieve==2.6
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
yahooquery==2.3.7
